[FIXED] Clean print/log code
