+++ FLASK/hstack/hstack/__init__.py
@@ -12,7 +12,7 @@ def create_app():
     app = Flask(__name__)
 
     app.config.from_pyfile('config.py')
-    app.config['SQLALCHEMY_DATABASE_URI'] = "mysql://root:hstackdbadmin@localhost/hstackdb"
+    app.config['SQLALCHEMY_DATABASE_URI'] = "mysql://root:csedbadmin@localhost/hstackdb"
     models.db.init_app(app)
     #DB = SQLAlchemy(app)
     #migrate = Migrate(app, DB)
