+++ FLASK/hstack/hstack/config.py
@@ -29,7 +29,7 @@ STT_API_KEY = {
 
 db = {
     'user' : 'root',
-    'password' : 'hstackdbadmin',
+    'password' : 'csedbadmin',
     'host' : 'localhost',
     'port' : 3306,
     'database' : 'hstackdb'
