+++ FLASK/hstack/uploadApi/config.py
@@ -30,10 +30,10 @@ STT_API_KEY = {
 
 db = {
     'user' : 'root',
-    'password' : 'hstackdbadmin',
+    'password' : 'csedbadmin',
     'host' : 'localhost',
     'port' : 3306,
-    'database' : 'hstackDB'
+    'database' : 'hstackdb'
 }
 
 DB = SQLAlchemy()
